requirements.txt
asgiref==3.8.1
Django==4.2.21
django-jalali==7.4.0
django-jalali-date==2.0.0
jalali_core==1.0.0
jdatetime==5.2.0
sqlparse==0.5.3
typing_extensions==4.13.2
